Refactor: Switch event extraction from OpenAI API to Google Studio API
I've replaced the OpenAI API integration with the Google Studio API for event detail extraction.

Key changes:
- I added `src/services/google-studio.ts` with an `extractEventDetails` function to interact with the Google Studio API.
- I updated `src/components/ExtractForm.tsx` to use the new service.
- I removed the old `src/services/chatgpt.ts`.
- I added placeholder environment variables `GOOGLE_STUDIO_URL` and `GOOGLE_STUDIO_API_KEY` in a new `.env` file.

**Important:**
- The `.env` file contains placeholder values for the Google Studio API endpoint and API key. You must replace these with actual credentials for the integration to work.
- The specific model and parameters for the Google Studio API in `src/services/google-studio.ts` are also placeholders and may need adjustment based on the chosen Google Studio model's capabilities.
- I didn't find any automated tests for the API integration, nor did I update any. Thorough manual testing is required to ensure the event extraction functionality works as expected.
