+++ src/services/google-studio.ts
@@ -1,0 +1,39 @@
+// This file will contain the logic for interacting with the Google Studio API.
+// Future implementations will include functions for making API calls,
+// handling responses, and managing authentication.
+
+export async function extractEventDetails(eventData: string): Promise<any> {
+  const response = await fetch(`${process.env.GOOGLE_STUDIO_URL}`, {
+    method: 'POST',
+    headers: {
+      Accept: 'application/json',
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${process.env.GOOGLE_STUDIO_API_KEY}`,
+    },
+    body: JSON.stringify({
+      // TODO: Replace with actual Google Studio model and parameters
+      model: 'google-studio-text-model', // Placeholder model name
+      prompt: `extract event details in google calendar api json format from this text: ${eventData}`,
+      temperature: 0, // Placeholder temperature
+      max_tokens: 100, // Placeholder max_tokens
+      top_p: 1, // Placeholder top_p
+      frequency_penalty: 0.0, // Placeholder frequency_penalty
+      presence_penalty: 0.0, // Placeholder presence_penalty
+      stop: ['\n'], // Placeholder stop sequences
+    }),
+  });
+
+  if (!response.ok) {
+    let errorData;
+    try {
+      errorData = await response.json();
+    } catch (e) {
+      // If the response is not JSON, use the status text
+      throw new Error(response.statusText);
+    }
+    // Attempt to extract a meaningful error message
+    throw new Error(JSON.stringify(errorData?.error?.message || errorData?.error?.code || errorData));
+  }
+
+  return response.json();
+}
